signed-unsigned updates

+++ libs/src/sio/sio.c
@@ -27,12 +27,12 @@
 /*
  * Stream write call: arguments same as those of write(2)
  */
-int Swrite( int fd, const char *addr, int nbytes )
+int Swrite( int fd, const char *addr, unsigned int nbytes )
 {
 	__sio_descriptor_t *dp ;
 	__sio_od_t *odp ;
-	int b_transferred ;
-	int b_avail ;
+	unsigned int b_transferred ;
+	unsigned int b_avail ;
 	int total_b_transferred ;
 	int b_written ;
 	int b_in_buffer ;
@@ -59,7 +59,7 @@ int Swrite( int fd, const char *addr, int nbytes )
 	b_in_buffer = odp->buf_end - odp->start ;
 	b_written = __sio_writef( odp, fd ) ;
 	if ( b_written != b_in_buffer )
-		return( (b_written >= nbytes) ? nbytes : b_written ) ;
+		return( (b_written >= (int)nbytes) ? (int)nbytes : b_written ) ;
 	
 	total_b_transferred = b_transferred ;
 	addr += b_transferred ;
@@ -80,7 +80,7 @@ int Swrite( int fd, const char *addr, int nbytes )
 		 * the buffer is full
 		 */
 		b_written = __sio_writef( odp, fd ) ;
-		if ( b_written != odp->buffer_size )
+		if ( b_written != (int)odp->buffer_size )
 		{
 			if ( b_written != SIO_ERR )
 			{
@@ -379,7 +379,7 @@ void __sio_memcopy( const char *from, char *to, int nbytes )
 
 #endif /* NEED_MEMCOPY */
 
-int sio_setup(int fd, __sio_descriptor_t **dp, int type)
+int sio_setup(int fd, __sio_descriptor_t **dp, unsigned int type)
 {
    if ( fd >= __sio_n_descriptors ) {
       if( Smorefds(fd) != 0 ) {
